Ensure LCD won't show weird characters due to electrical noise
While the pump is on, it drains current from the source as the coils
switch fastly, which creates a high noisy on the circuitry of the LCD,
causing it to behave unexpectedly.
This change ensures that even if this happens, the LCD will be back in
shape just in a cycle. The use of .1uF ceramic capacitors help a lot
(one in parallel to each pump, as close as possible), and one VCC/GND as
close as possible to the LCD.