Separate factor from time set up by user
The Clock normally is set up following the workflow:
- set up hour
- set up minute
- wait for a couple of hours
- tune factor

However there is an error in this way of proceeding. This is because the
cyclesFromT0 is filled when hour/minutes are set up taking into account
the wrong initial factor. When the factor is changed afterwards the
initial cyclesFromT0 does not make sense anymore.
This change separates the time initially set up by the user from the
increments added by the system (which are the only amount impacted by
the factor).