Try new mechanism for serving WDT interrupts
Interrupts are sort of queued.
If 2 interrupts happened, a counter will count 2, and a Module cycle
will be executed telling a WDT interrupt happened, and then decreasing
by one the counter. If the counter is greater than 0 the UC will not
sleep so that the WDT interrupt pending can be served as soon as
possible.
This change is based on the assumption that increasing a volatile char
is atomic (as it will be modified by the ISR increasing it and by the
loop decreasing it).